24 07 29 수업
Tessellation LOD 구현
Texture 2D Array 구현 중

+++ Project/Engine/struct.fx
@@ -129,7 +129,7 @@ struct tParticleModule
 struct tRaycastOut
 {
     float2  Location;
-    float   Distance;
+    uint    Distance;
     int     Success;
 };
